Add files via upload
the individual-level survey level mortality dataset of Bihar (a state in India) containing information regarding the various deaths of individuals from the year 2007 to 2011 has been analyzed. A logistic regression prediction model has been developed which analyses the deaths in Bihar state from 2007 to 2011 and predicts whether the death has been registered or not. This prediction model has been tested on the test set and 18071 deaths registration status has been correctly predicted out of the total 19893 deaths achieving an accuracy of 90.8 percent
•	Technology:  R, RStudio